# public/docs/openapi.yaml
openapi: 3.0.3
info:
    title: 'OnesBlog API Documentation'
    description: ''
    version: 1.0.0
servers:
    -
        url: 'http://ones-blog-api.test'
paths:
...
    '/api/posts/{post}':
        get:
            summary: 'Display the specified post.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200, when post displayed.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: { created_at: '{post-created-at}', user: { id: '{user-id}', created_at: '{user-created-at}', updated_at: '{user-updated-at}', name: '{user-name}', email: '{user-email}', email_verified_at: '{user-email-verified-at}', login_type_id: '{user-login-type-id}' }, location_id: '{post-location-id}', title: '{post-title}', content: '{post-content}', published_at: '{post-published-at}', slug: '{post-slug}', images: '{post-images}' }
                                properties:
                                    data: { type: object, example: { created_at: '{post-created-at}', user: { id: '{user-id}', created_at: '{user-created-at}', updated_at: '{user-updated-at}', name: '{user-name}', email: '{user-email}', email_verified_at: '{user-email-verified-at}', login_type_id: '{user-login-type-id}' }, location_id: '{post-location-id}', title: '{post-title}', content: '{post-content}', published_at: '{post-published-at}', slug: '{post-slug}', images: '{post-images}' } }
                404:
                    description: '404, when post not found, inactive or unpublished.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: 'Post(ID:{post-id}) is not found.'
                                properties:
                                    data: { type: string, example: 'Post(ID:{post-id}) is not found.' }
            tags:
                - '06. Posts'
            security: []
        put:
            summary: 'Update the specified post in storage.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: multipart/form-data
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Authorization
                    description: ''
                    example: 'Bearer {personal-access-token}'
                    schema:
                        type: string
            responses:
                200:
                    description: '200, when post updated.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: { created_at: '{post-created-at}', user: { id: '{user-id}', created_at: '{user-created-at}', updated_at: '{user-updated-at}', name: '{user-name}', email: '{user-email}', email_verified_at: '{user-email-verified-at}', login_type_id: '{user-login-type-id}' }, location_id: '{post-location-id}', title: '{post-title}', content: '{post-content}', published_at: '{post-published-at}', slug: '{post-slug}', images: '{post-images}' }
                                properties:
                                    data: { type: object, example: { created_at: '{post-created-at}', user: { id: '{user-id}', created_at: '{user-created-at}', updated_at: '{user-updated-at}', name: '{user-name}', email: '{user-email}', email_verified_at: '{user-email-verified-at}', login_type_id: '{user-login-type-id}' }, location_id: '{post-location-id}', title: '{post-title}', content: '{post-content}', published_at: '{post-published-at}', slug: '{post-slug}', images: '{post-images}' } }
                401:
                    description: '401, without personal access token.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: Unauthenticated.
                                properties:
                                    data: { type: string, example: Unauthenticated. }
                404:
                    description: '404, when post not found.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: 'Post(ID:{post-id}) is not found.'
                                properties:
                                    data: { type: string, example: 'Post(ID:{post-id}) is not found.' }
                422:
                    description: '422, when any validation failed.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: { title: ['{validation-error-message}'], published_at: ['{validation-error-message}'], active: ['{validation-error-message}'] }
                                properties:
                                    data: { type: object, example: { title: ['{validation-error-message}'], published_at: ['{validation-error-message}'], active: ['{validation-error-message}'] } }
            tags:
                - '06. Posts'
            requestBody:
                required: false
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                    description: 'The title of the post.'
                                    example: 0724Post
                                content:
                                    type: string
                                    description: 'The content of the post.'
                                    example: 0724Test
                                published_at:
                                    type: string
                                    description: 'The published time of the post.'
                                    example: '20220724'
                                active:
                                    type: boolean
                                    description: 'The state of the post.'
                                    example: true
                                images:
                                    type: array
                                    description: 'The images of the post.'
                                    example: [null, null]
                                    items: { type: file }
                                _method:
                                    type: string
                                    description: 'Required if the <code><b>images</b></code> of the post are uploaded, must be <b>PUT</b> and request method must be <small class="badge badge-black">POST</small>.'
                                    example: PUT
        delete:
            summary: 'Remove the specified post from storage.'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Authorization
                    description: ''
                    example: 'Bearer {personal-access-token}'
                    schema:
                        type: string
            responses:
                200:
                    description: '200, when post deleted.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: Success
                                properties:
                                    data: { type: string, example: Success }
                401:
                    description: '401, without personal access token.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: Unauthenticated.
                                properties:
                                    data: { type: string, example: Unauthenticated. }
                404:
                    description: '404, when post not found.'
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: 'Post(ID:{post-id}) is not found.'
                                properties:
                                    data: { type: string, example: 'Post(ID:{post-id}) is not found.' }
            tags:
                - '06. Posts'
        parameters:
            -
                in: path
                name: post
                description: 'The id of the post.'
                example: 108
                required: true
                schema:
                    type: integer
...